requirements.txt
APScheduler==3.11.2
python-dotenv==1.2.2
tzlocal==5.3.1
<<<<<<< Automation-role
=======
python-telegram-bot==21.9
httpx==0.28.1
>>>>>>> local
